requirements.txt
fastapi
uvicorn
python-multipart
httpx
google-generativeai
openai-whisper
ffmpeg-python
torch
python-dotenv
